refactor: split trade contract, debug helpers, and tree builders (#3878)
Extract Market row contract keys, Market tab section assembly, debug
command helper parts, and naval/military tree-builder model/assembly
slices to keep hosts under file-size targets without API changes.
